// src/shared/enums/message.status.enum.ts
export enum MessageStatus {
	ERRORCREATESENSORS = 'Error guardando sensor',
	ERRORUPDATESENSORS = 'Error actualizando sensor',
	ERRORSHOWSENSORS = 'Error consultado sensor',
	ERRORDELETESENSORS = 'Error eliminando sensor',
	EXITOCREATESENSORS = 'Sensor guardado con exito',
	EXITOUPDATESENSORS = 'Sensor actualizado con exito',
	EXITOSHOWSENSORS = 'Sensor consultado con exito',
	EXITODELETESENSORS = 'Sensor eliminado con exito',
	ERRORCREATEENTITIES = 'Error guardando entidad',
	ERRORUPDATEENTITIES = 'Error actualizando entidad',
	ERRORSHOWENTITIES = 'Error consultando entidad',
	ERRORDELETEENTITIES = 'Error eliminando entidad',
	EXITOCREATEENTITIES = 'EXITO guardando entidad',
	EXITOUPDATEENTITIES = 'EXITO actualizando entidad',
	EXITOSHOWENTITIES = 'EXITO consultando entidad',
	EXITODELETEENTITIES = 'EXITO eliminando entidad',

	ERRORCREATEHISTORIES = 'Error guardando registro',
	ERRORUPDATEHISTORIES = 'Error actualizando registro',
	ERRORSHOWHISTORIES = 'Error consultando registro',
	ERRORDELETEHISTORIES = 'Error eliminando registro',
	EXITOCREATEHISTORIES = 'EXITO guardando registro',
	EXITOUPDATEHISTORIES = 'EXITO actualizando registro',
	EXITOSHOWHISTORIES = 'EXITO consultando registro',
	EXITODELETEHISTORIES = 'EXITO eliminando registro',

	ERRORCREATESUGGESTIONS = 'Error guardando sugerencia',
	ERRORUPDATESUGGESTIONS = 'Error actualizando sugerencia',
	ERRORSHOWSUGGESTIONS = 'Error consultando sugerencia',
	ERRORDELETESUGGESTIONS = 'Error eliminando sugerencia',
	EXITOCREATESUGGESTIONS = 'EXITO guardando sugerencia',
	EXITOUPDATESUGGESTIONS = 'EXITO actualizando sugerencia',
	EXITOSHOWSUGGESTIONS = 'EXITO consultando sugerencia',
	EXITODELETESUGGESTIONS = 'EXITO eliminando sugerencia'
}